@@ -1,16 +1,15 @@
 package ru.api;
 
-import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLn;
-import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLnImplService;
-import ru.ibs.fss.ln.ws.fileoperationsln.SOAPException_Exception;
-import ru.my.helpers_operations.GlobalVariables;
-
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 
+import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLn;
+import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLnImplService;
+import ru.ibs.fss.ln.ws.fileoperationsln.SOAPException_Exception;
+import ru.my.utils.GlobalVariables;
 
 @Path("/import")
 public class Import {
