@@ -1,11 +1,12 @@
 package ru.my.entities;
 
-import javax.xml.bind.annotation.*;
 import java.util.List;
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlAttribute;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlRootElement;
 
-/**
- * Created by rkurbanov on 05.12.2016.
- */
 @XmlRootElement(name = "ROWSET")
 @XmlAccessorType(XmlAccessType.FIELD)
 public class ROWSET {
@@ -28,45 +29,59 @@ public class ROWSET {
     public List<ROW> getRow() {
         return row;
     }
+
     public void setRow(List<ROW> row) {
         this.row = row;
     }
+
     public String getAuthor() {
         return author;
     }
+
     public void setAuthor(String author) {
         this.author = author;
     }
+
     public String getVersion() {
         return version;
     }
+
     public void setVersion(String version) {
         this.version = version;
     }
+
     public String getSoftware() {
         return software;
     }
+
     public void setSoftware(String software) {
         this.software = software;
     }
+
     public String getVersionSoftware() {
         return versionSoftware;
     }
+
     public void setVersionSoftware(String versionSoftware) {
         this.versionSoftware = versionSoftware;
     }
+
     public String getPhone() {
         return phone;
     }
+
     public void setPhone(String phone) {
         this.phone = phone;
     }
+
     public String getEmail() {
         return email;
     }
+
     public void setEmail(String email) {
         this.email = email;
     }
+
     @Override
     public String toString() {
         String res = "ROWSET{" +
