@@ -1,20 +1,19 @@
 package ru.my.servlets;
-import javax.servlet.*;
+
 import java.io.File;
+import javax.servlet.ServletContextEvent;
+import javax.servlet.ServletContextListener;
+
 import org.apache.log4j.PropertyConfigurator;
-
-/**
- * Created by rkurbanov on 23.05.2017.
- */
-
 
 public class Log4jInit implements ServletContextListener {
 
     public void contextInitialized(ServletContextEvent event) {
-        String homeDir=event.getServletContext().getRealPath("/");
-        File propertiesFile=new File(homeDir,"WEB-INF/log4j.properties");
+        String homeDir = event.getServletContext().getRealPath("/");
+        File propertiesFile = new File(homeDir, "WEB-INF/log4j.properties");
         PropertyConfigurator.configure(propertiesFile.toString());
     }
 
-    public void contextDestroyed(ServletContextEvent event) {}
+    public void contextDestroyed(ServletContextEvent event) {
+    }
 }
