@@ -1,33 +1,30 @@
 package ru.my.servlets;
 
-import org.apache.log4j.Logger;
-import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLn;
-import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLnImplService;
-import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLnUserGetExistingLNNumRangeOut;
-import ru.ibs.fss.ln.ws.fileoperationsln.LnNumList;
-import ru.my.helpers_operations.GlobalVariables;
-
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.util.List;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.List;
 
-/**
- * Created by rkurbanov on 30.03.2018.
- */
+import org.apache.log4j.Logger;
+
+import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLn;
+import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLnImplService;
+import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLnUserGetExistingLNNumRangeOut;
+import ru.ibs.fss.ln.ws.fileoperationsln.LnNumList;
+import ru.my.utils.GlobalVariables;
+
 @WebServlet("/sExistingLNNumRange")
 public class sExistingLNNumRange extends HttpServlet {
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
 
-        Logger logger= Logger.getLogger("simple");
+        Logger logger = Logger.getLogger("simple");
         logger.info("1) sExistingLNNumRange");
         response.setContentType("text/html ;charset=UTF-8");
 
         String ogrn = request.getParameter("ogrn");
-
 
         PrintWriter out = response.getWriter();
         out.println("<html>");
@@ -37,8 +34,7 @@ public class sExistingLNNumRange extends HttpServlet {
                 "  <meta http-equiv=\"Content-Type\" content=\"text/html; charset=utf-8\"> </head>");
 
         out.println("<body>");
-        out.print("<H1> ogrn="+ogrn+"</H1>");
-
+        out.print("<H1> ogrn=" + ogrn + "</H1>");
 
         System.setProperty("javax.net.ssl.trustStore", GlobalVariables.pathandnameSSL);//КОНФ
         System.setProperty("javax.net.ssl.trustStorePassword", GlobalVariables.passwordSSL);
@@ -46,27 +42,23 @@ public class sExistingLNNumRange extends HttpServlet {
         FileOperationsLn start = service.getFileOperationsLnPort();
 
         try {
-
             FileOperationsLnUserGetExistingLNNumRangeOut rangeOut = start.getExistingLNNumRange(ogrn);
+            out.print("<H1> Инфо=" + rangeOut.getINFO() + "</H1>");
+            out.print("<H1> Сообщение=" + rangeOut.getMESS() + "</H1>");
+            out.print("<H1> Статус=" + rangeOut.getSTATUS() + "</H1>");
 
-            out.print("<H1> Инфо="+rangeOut.getINFO()+"</H1>");
-            out.print("<H1> Сообщение="+rangeOut.getMESS()+"</H1>");
-            out.print("<H1> Статус="+rangeOut.getSTATUS()+"</H1>");
-
-            if(rangeOut.getDATA()!=null){
+            if (rangeOut.getDATA() != null) {
                 LnNumList s = rangeOut.getDATA();
                 List<String> lns = s.getLNNum();
-                for (String ln:lns){
-                    out.print("<H1> eln# "+ln+"</H1>");
+                for (String ln : lns) {
+                    out.print("<H1> eln# " + ln + "</H1>");
                 }
-            }else {
+            } else {
                 out.print("<H1> Нет данных по неиспользованным номерам </H1>");
             }
-
-        }catch(Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
-
         out.println("</body>");
         out.println("</html>");
     }
