@@ -1,31 +1,27 @@
 package ru.my.servlets;
 
-import org.apache.log4j.Logger;
-import ru.ibs.fss.ln.ws.fileoperationsln.*;
-import ru.my.helpers_operations.GlobalVariables;
-
+import java.io.IOException;
+import java.io.PrintWriter;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.PrintWriter;
 
-import static ru.my.helpers_operations.SQL.SQL_UpdIns;
-import static ru.my.helpers_operations.StoredQuery.SaveNumber;
+import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLn;
+import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLnImplService;
+import ru.ibs.fss.ln.ws.fileoperationsln.FileOperationsLnUserGetNewLNNumOut;
+import ru.ibs.fss.ln.ws.fileoperationsln.SOAPException_Exception;
+import ru.my.utils.GlobalVariables;
 
-/**
- * Created by rkurbanov on 28.06.2017.
- */
-//C:\Program Files\Java\jre1.8.0_131\bin\client\jvm.dll
+import static ru.my.utils.SQL.sqlUpdIns;
+import static ru.my.utils.StoredQuery.SaveNumber;
+
 @WebServlet("/sNewLnNum")
 public class sNewLnNum extends HttpServlet {
+
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        Logger logger=Logger.getLogger("simple");
-        logger.info("1) NewLnNum");
         response.setContentType("text/html ;charset=UTF-8");
         String ogrn = request.getParameter("ogrn");
-
         GlobalVariables.requestParam = ogrn;
 
         PrintWriter out = response.getWriter();
@@ -36,30 +32,26 @@ public class sNewLnNum extends HttpServlet {
                 "  <meta http-equiv=\"Content-Type\" content=\"text/html; charset=utf-8\"> </head>");
 
         out.println("<body>");
-        out.print("<H1> ogrn="+ogrn+"</H1>");
+        out.print("<H1> ogrn=" + ogrn + "</H1>");
 
-        System.setProperty("javax.net.ssl.trustStore",GlobalVariables.pathandnameSSL);//КОНФ
-        logger.info("1) trustStore 1");
+        System.setProperty("javax.net.ssl.trustStore", GlobalVariables.pathandnameSSL);
         System.setProperty("javax.net.ssl.trustStorePassword", GlobalVariables.passwordSSL);
-        logger.info("1) trustStore 2");
 
-        FileOperationsLnImplService service=null;
-        FileOperationsLn start=null;
-        try{
-            service = new  FileOperationsLnImplService();
-        logger.info("1) serv 2");
-         start = service.getFileOperationsLnPort();
-        logger.info("1) send 2");}catch (Exception e) {
+        FileOperationsLnImplService service = null;
+        FileOperationsLn start = null;
+        try {
+            service = new FileOperationsLnImplService();
+            start = service.getFileOperationsLnPort();
+        } catch (Exception e) {
             e.printStackTrace();
         }
 
-
         try {
             FileOperationsLnUserGetNewLNNumOut Num = start.getNewLNNum(ogrn);//getNewLNNumRange(ogrn,count);
-            out.println("<H1>"+Num.getMESS()+"</H1>");
+            out.println("<H1>" + Num.getMESS() + "</H1>");
             out.println("<H1> Получен номер: </H1>");
-            out.println("<H1>1)"+Num.getDATA()+"</H1>");
-            SQL_UpdIns(SaveNumber(Num.getDATA()));
+            out.println("<H1>1)" + Num.getDATA() + "</H1>");
+            sqlUpdIns(SaveNumber(Num.getDATA()));
 
         } catch (SOAPException_Exception e) {
             e.printStackTrace();
