@@ -1,31 +1,26 @@
 package ru.my.servlets;
 
-import org.apache.log4j.Logger;
-import ru.ibs.fss.ln.ws.fileoperationsln.*;
-import ru.my.helpers_operations.GlobalVariables;
-import ru.my.helpers_operations.SQL;
-import ru.my.helpers_operations.StoredQuery;
-
-import javax.servlet.ServletException;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.List;
 
-import static ru.my.helpers_operations.GlobalVariables.*;
+import org.apache.log4j.Logger;
 
-//Created by rashgild on 25.05.2017.
+import ru.ibs.fss.ln.ws.fileoperationsln.*;
+import ru.my.utils.GlobalVariables;
+import ru.my.utils.SQL;
+import ru.my.utils.StoredQuery;
+
+import static ru.my.utils.GlobalVariables.*;
 
 @WebServlet("/SetLnData")
 public class sPrParseFileLnLpu extends HttpServlet {
 
-
-    public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
-
+    public void doGet(HttpServletRequest request, HttpServletResponse response) {
         try {
             long start = System.currentTimeMillis();
             response.setContentType("text/html ;charset=UTF-8");
@@ -33,20 +28,20 @@ public class sPrParseFileLnLpu extends HttpServlet {
             GlobalVariables.requestParam = id;
             request.setAttribute("id", id);
 
-            boolean f =getLnHash(id);
+            boolean f = getLnHash(id);
             request.setAttribute("snils", f);
-            if(f) {
+            if (f) {
                 WSResult result = setRequest();
                 List<INFO.ROWSET.ROW> rows = result.getINFO().getROWSET().getROW();
                 StringBuilder saveResult = new StringBuilder();
-                String state = "", hash = "",status="";
+                String state = "", hash = "", status = "";
 
                 if (rows != null && rows.size() > 0) {
                     for (INFO.ROWSET.ROW row : rows) {
 
                         state = row.getLNSTATE();
                         hash = row.getLNHASH();
-                        if(state.equals("") && hash.equals("")){
+                        if (state.equals("") && hash.equals("")) {
                             hash = GlobalVariables.hash;
                             state = GlobalVariables.state;
                         }
@@ -57,44 +52,46 @@ public class sPrParseFileLnLpu extends HttpServlet {
 
                         saveResult.append(result.getMESS());
                         try {
-                            if (row.getERRORS()!=null && row.getERRORS().getERROR().size() > 0) {
+                            if (row.getERRORS() != null && row.getERRORS().getERROR().size() > 0) {
                                 List<INFO.ROWSET.ROW.ERRORS.ERROR> errors = row.getERRORS().getERROR();
                                 request.setAttribute("errors", errors);
                                 for (INFO.ROWSET.ROW.ERRORS.ERROR errs : errors) {
                                     saveResult.append(":".concat(errs.getERRMESS()));
                                 }
                             }
-                        }catch (Exception e){e.printStackTrace();}
+                        } catch (Exception e) {
+                            e.printStackTrace();
+                        }
                     }
                 }
 
                 request.setAttribute("time", ((System.currentTimeMillis() - start) / 1000));
                 if (state != null && !state.equals("") || hash != null && !hash.equals("")) {
-                    SQL.SQL_UpdIns(StoredQuery.SaveStatusAndHash(state, hash, GlobalVariables.t_ELN));
-                    if(status.equals("1")){
-                        SQL.SQL_UpdIns(StoredQuery.updateDisRecord(id));
+                    SQL.sqlUpdIns(StoredQuery.SaveStatusAndHash(state, hash, GlobalVariables.t_ELN));
+                    if (status.equals("1")) {
+                        SQL.sqlUpdIns(StoredQuery.updateDisRecord(id));
                     }
                 }
-                SQL.SaveInBD(saveResult.toString(), result.getSTATUS());
+                SQL.saveInBaseDate(saveResult.toString(), result.getSTATUS());
             }
             request.getRequestDispatcher("/WEB-INF/prParseFileLnLpu.jsp").forward(request, response);
 
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
     }
 
-    private static WSResult setRequest(){
-        System.setProperty("javax.net.ssl.trustStore",pathandnameSSL);
+    private static WSResult setRequest() {
+        System.setProperty("javax.net.ssl.trustStore", pathandnameSSL);
         System.setProperty("javax.net.ssl.trustStorePassword", passwordSSL);
-        FileOperationsLnImplService service = new  FileOperationsLnImplService();
+        FileOperationsLnImplService service = new FileOperationsLnImplService();
         FileOperationsLn start = service.getFileOperationsLnPort();
         ROWSET rowset = new ROWSET();
         PrParseFilelnlpuElement prParseFilelnlpuElement = new PrParseFilelnlpuElement();
-        PrParseFilelnlpuElement.PXmlFile pXmlFile= new PrParseFilelnlpuElement.PXmlFile();
+        PrParseFilelnlpuElement.PXmlFile pXmlFile = new PrParseFilelnlpuElement.PXmlFile();
         pXmlFile.setROWSET(rowset);
         prParseFilelnlpuElement.setPXmlFile(pXmlFile);
-        WSResult result =null;
+        WSResult result = null;
         try {
             result = start.prParseFilelnlpu(prParseFilelnlpuElement);
             return result;
@@ -104,34 +101,55 @@ public class sPrParseFileLnLpu extends HttpServlet {
         return result;
     }
 
-    private static boolean getLnHash(String id){
-        Logger logger=Logger.getLogger("");
+    private static boolean getLnHash(String id) {
+        Logger logger = Logger.getLogger("");
         logger.info("Get Hash");
-        String snils="",eln="";
-        ResultSet resultSet = SQL.select(StoredQuery.getLNandSnils(id));
+        String snils = "", eln = "";
+        ResultSet rs = SQL.select(StoredQuery.getLnHash(id));
+        String hash = null;
         try {
-            while (resultSet.next()) {
-                snils = resultSet.getString("snils");
-                eln =resultSet.getString("ln_code");
+            while (rs.next()){
+                hash = rs.getString("lasthash");
+                System.out.println(GlobalVariables.hash + ">>>>from base hash");
             }
-        } catch (SQLException e) {e.printStackTrace();}
-
-        if(snils==null || snils.length()<=0) return false;
-        snils=snils.replace(" ","").replace("-","");
-        logger.info("Snils: "+snils);
-        logger.info("Eln: "+eln);
-        try {
-            System.setProperty("javax.net.ssl.trustStore",GlobalVariables.pathandnameSSL);
-            System.setProperty("javax.net.ssl.trustStorePassword", GlobalVariables.passwordSSL);
-            FileOperationsLnImplService service = new  FileOperationsLnImplService();
-            FileOperationsLn start = service.getFileOperationsLnPort();
-            FileOperationsLnUserGetLNDataOut fileOperationsLnUserGetLNDataOut = start.getLNData(GlobalVariables.ogrnMo,eln,snils);
-            GlobalVariables.hash  = fileOperationsLnUserGetLNDataOut.getDATA().getOUTROWSET().getROW().get(0).getLNHASH();
-            GlobalVariables.state  = fileOperationsLnUserGetLNDataOut.getDATA().getOUTROWSET().getROW().get(0).getLNSTATE();
-
-        } catch (Exception e) {
+        } catch (SQLException e) {
             e.printStackTrace();
         }
-        return true;
+
+        if(hash==null) {
+            ResultSet resultSet = SQL.select(StoredQuery.getLNandSnils(id));
+            try {
+                while (resultSet.next()) {
+                    snils = resultSet.getString("snils");
+                    eln = resultSet.getString("ln_code");
+                }
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+
+            if (snils == null || snils.length() <= 0) return false;
+            snils = snils.replace(" ", "").replace("-", "");
+            logger.info("Snils: " + snils);
+            logger.info("Eln: " + eln);
+            try {
+                System.setProperty("javax.net.ssl.trustStore", GlobalVariables.pathandnameSSL);
+                System.setProperty("javax.net.ssl.trustStorePassword", GlobalVariables.passwordSSL);
+                FileOperationsLnImplService service = new FileOperationsLnImplService();
+                FileOperationsLn start = service.getFileOperationsLnPort();
+                FileOperationsLnUserGetLNDataOut fileOperationsLnUserGetLNDataOut = start.getLNData(GlobalVariables.ogrnMo, eln, snils);
+                for (ROW row : fileOperationsLnUserGetLNDataOut.getDATA().getOUTROWSET().getROW()) {
+                    GlobalVariables.hash = row.getLNHASH();
+                    System.out.println(GlobalVariables.hash + ">>>>has");
+                    GlobalVariables.state = row.getLNSTATE();
+                }
+
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            return true;
+        } else {
+            GlobalVariables.hash = hash;
+            return true;
+        }
     }
 }
