@@ -1,23 +1,18 @@
-package ru.my.helpers_operations;
+package ru.my.utils;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.URL;
 import java.net.URLConnection;
+import java.nio.charset.StandardCharsets;
 import java.util.Locale;
 import java.util.PropertyResourceBundle;
 import java.util.ResourceBundle;
 
-/**
- * Created by rkurbanov on 08.06.2017.
- */
 public class UTF8Control extends ResourceBundle.Control {
     public ResourceBundle newBundle
-            (String baseName, Locale locale, String format, ClassLoader loader, boolean reload)
-            throws IllegalAccessException, InstantiationException, IOException
-    {
-        // The below is a copy of the default implementation.
+            (String baseName, Locale locale, String format, ClassLoader loader, boolean reload) throws IOException {
         String bundleName = toBundleName(baseName, locale);
         String resourceName = toResourceName(bundleName, "properties");
         ResourceBundle bundle = null;
@@ -36,8 +31,7 @@ public class UTF8Control extends ResourceBundle.Control {
         }
         if (stream != null) {
             try {
-                // Only this line is changed to make it to read properties files as UTF-8.
-                bundle = new PropertyResourceBundle(new InputStreamReader(stream, "UTF-8"));
+                bundle = new PropertyResourceBundle(new InputStreamReader(stream, StandardCharsets.UTF_8));
             } finally {
                 stream.close();
             }
