@@ -1,23 +1,36 @@
-package ru.my.helpers_operations;
+package ru.my.utils;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.io.OutputStreamWriter;
+import java.io.StringWriter;
+import java.io.Writer;
+import javax.xml.soap.MessageFactory;
+import javax.xml.soap.SOAPException;
+import javax.xml.soap.SOAPMessage;
+import javax.xml.transform.OutputKeys;
+import javax.xml.transform.Transformer;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
 
 import org.apache.xml.security.c14n.Canonicalizer;
 import org.w3c.dom.Document;
 
-import javax.xml.soap.MessageFactory;
-import javax.xml.soap.SOAPException;
-import javax.xml.soap.SOAPMessage;
-import javax.xml.transform.*;
-import javax.xml.transform.dom.DOMResult;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
-import java.io.*;
+public class XmlUtils {
 
-/**
- * Created by rkurbanov on 22.05.2017.
- */
-public class WorkWithXML {
-
-    public static String DocToString(Document doc) {
+    /**
+     * w3c.dom.document to string.
+     *
+     * @param doc converting
+     * @return String
+     */
+    public static String docToString(Document doc) {
         try {
             StringWriter sw = new StringWriter();
             TransformerFactory tf = TransformerFactory.newInstance();
@@ -38,27 +51,22 @@ public class WorkWithXML {
         transformer.transform(new DOMSource(doc), new StreamResult(out));
     }
 
-    public byte[] documentToByte(Document document)
-    {
+    public byte[] documentToByte(Document document) {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         org.apache.xml.security.utils.XMLUtils.outputDOM(document, baos, true);
         return baos.toByteArray();
     }
 
     /**
-     * Сохранение Soap-сообщения в файл
-     * @param FileName Имя сохраняемого файла
+     * Save Soap-message to file.
+     *
+     * @param fileName    Имя сохраняемого файла
      * @param soapMessage Сообщение, которое требуется сохранить
-     * @throws Exception e,
-     * @throws IOException e
      */
-    public static void SaveSOAPToXML( String FileName, SOAPMessage soapMessage)
-            throws IOException, SOAPException {
-
-        String strMsg = SoapMessageToString(soapMessage);//new String(out2.toByteArray());//.getBytes("UTF-16");
+    public static void saveSoapToXml(String fileName, SOAPMessage soapMessage) {
+        String strMsg = soapMessageToString(soapMessage);
         try {
-
-            Writer w = new OutputStreamWriter(new FileOutputStream(GlobalVariables.pathtosave+FileName), "UTF-8");
+            Writer w = new OutputStreamWriter(new FileOutputStream(GlobalVariables.pathtosave + fileName), "UTF-8");
             w.write(strMsg);
             w.flush();
             w.close();
@@ -69,11 +77,12 @@ public class WorkWithXML {
     }
 
     /**
-     * Конверт из w3.dom.Document в Soap-сообщение
-     * @param doc конвертируемый документ
+     * Convert from w3.dom.Document to SOAP-message.
+     *
+     * @param doc converting document
      * @return SOAPMessage
      */
-    public static SOAPMessage DocToSOAP(Document doc) throws Exception {
+    public static SOAPMessage docToSoap(Document doc) throws Exception {
         Canonicalizer c14n =
                 Canonicalizer.getInstance(Canonicalizer.ALGO_ID_C14N_WITH_COMMENTS);
         byte[] canonicalMessage = c14n.canonicalizeSubtree(doc);
@@ -82,44 +91,36 @@ public class WorkWithXML {
         return factory.createMessage(null, in);
     }
 
-    public static Document soapToDoc(SOAPMessage soapMsg)
-            throws TransformerException, SOAPException, IOException {
-      /*  MyByteArrayOutputStream baos = new MyByteArrayOutputStream();
-        soapMsg.writeTo(baos);
-        ByteArrayInputStream bais = baos.getByteArrayInputStream();
-
-        DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
-        dbf.setNamespaceAware(true);
-        DocumentBuilder db = dbf.newDocumentBuilder();
-        Document doc = db.parse(bais);
-        return(doc);*/
-
-        Document doc = null;
-      return doc;
-    }
-
-    public static String SoapMessageToString(SOAPMessage soapMessage)
-    {
+    /**
+     * SOAP-message to sting.
+     *
+     * @param soapMessage to converting
+     * @return String
+     */
+    public static String soapMessageToString(SOAPMessage soapMessage) {
         try {
             ByteArrayOutputStream stream = new ByteArrayOutputStream();
             soapMessage.writeTo(stream);
             String message = new String(stream.toByteArray(), "utf-8");
             return message;
-        }
-        catch (SOAPException | IOException e) {
+        } catch (SOAPException | IOException e) {
             e.printStackTrace();
         }
         return null;
     }
 
-    public static SOAPMessage stringToSoap(String soap){
-
-        SOAPMessage soapMessage=null;
-
+    /**
+     * String to SOAP-message,
+     *
+     * @param soap string to convert
+     * @return SOAP-message
+     */
+    public static SOAPMessage stringToSoap(String soap) {
+        SOAPMessage soapMessage = null;
         try {
             InputStream is = new ByteArrayInputStream(soap.getBytes());
             soapMessage = MessageFactory.newInstance().createMessage(null, is);
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
         return soapMessage;
